+++ CHANGELOG.md
@@ -17,6 +17,7 @@
  - add docker configurations
  - add examples that can run in docker
  - add if statement
+ - add explicit custom buttons
 
 ## v1.0 (25/03/2023)